requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-openai
sentence-transformers
faiss-cpu
openai
azure-core
azure-storage-blob
langsmith
python-multipart
